backend/requirements.txt
# llama-index
llama-index
psycopg2-binary

# neon pgvector
asyncpg
llama-index-vector-stores-postgres

# langchain
langchain
langchain-openai
langgraph

# for serper
langchain-community
google-search-results

# langsmith
langsmith
openai

# checkpointer postgres
langgraph-checkpoint-postgres
psycopg-pool
psycopg

fastapi
uvicorn
pypdf
requests
python-dotenv
